refactor: internal/storageパッケージとFirestore移行ツールを削除
- UserKeyをmodel、CSV生成をpipelineに移動
- cloud_storage.go（GCS操作）を削除
- cmd/migrate-to-firestore/（移行済み）を削除
- go.modからcloud.google.com/go/storage依存を削除
